requirements.txt
pandas>=2.2
openpyxl>=3.1
docxtpl>=0.16
python-slugify>=8.0
typer>=0.12
rich>=13.7
